Refactor the structure of stories & tests
