+++ src/components/button/tests/expected/chrome-button-large.png
