+++ src/components/button/tests/expected/chrome-button-medium.png
