+++ src/components/button/tests/expected/chrome-button-small.png
