+++ src/components/button/tests/index.spec.ts
@@ -1,4 +1,4 @@
-import { VisualRegressionTest } from '../../../lib/test/visual-regression-test';
+import { VisualRegressionTest } from '../../../../lib/test/visual-regression-test';
 import * as style from '../style.less';
 
 new VisualRegressionTest( __dirname, style.container).run();
