+++ src/components/text/tests/expected/chrome-text-with-black-background.png
