+++ src/components/text/tests/expected/chrome-text-with-red-background.png
