+++ tsconfig.json
@@ -14,11 +14,6 @@
     "allowJs": false,
     "jsx": "react",
     "moduleResolution": "node",
-    "rootDirs": [
-      ".",
-      "src",
-      "stories"
-    ],
     "baseUrl": ".",
     "forceConsistentCasingInFileNames": true,
     "noImplicitReturns": true,
@@ -34,7 +29,7 @@
     "types": ["@wdio/mocha-framework", "node", "@wdio/sync"]
   },
   "include": [
-    "src/**/*", ".storybook/stories.tsx", "typings", "*.d.ts"
+    "", "src/**/*", ".storybook/stories.tsx", "typings", "*.d.ts"
   ],
   "exclude": [
     "node_modules",
